move json formatter into format.go
